CV update logic moved to frontend login profile page, WIP
