+++ components/controllers/auth/alive.js
@@ -1,6 +1,4 @@
 const status = (req, res) => {
-  console.log('Ping pong');
-
   res.json({ ping: 'pong' });
 };
