+++ components/controllers/auth/login.js
@@ -9,8 +9,6 @@ const JWT_SECRET = process.env.JWT_SECRET;
 const JWT_EXPIRATION = '1h';
 
 const login = async (req, res) => {
-  console.log('Login request');
-
   const { email, password } = req.body;
   db.all('SELECT * FROM users where email = ?', [email], async (err, rows) => {
     if (err) {
