+++ components/controllers/cv/getSkills.js
@@ -1,7 +1,6 @@
 import db from '../../database/database.js';
 
 const getSkills = (req, res) => {
-  console.log('New skills get request');
   db.all('SELECT * FROM skills', [], (err, rows) => {
     if (err) {
       res.status(500).json({ error: err.message });
