+++ components/controllers/users/createUser.js
@@ -2,8 +2,6 @@ import db from '../../database/database.js';
 import { hashPassword } from '../../helpers/encryption.js';
 
 const createUser = async (req, res) => {
-  console.log("Register request");
-  
   let hashedPassword;
   const { firstName, lastName, email, password } = req.body;
   if (password) {
