+++ main.js
@@ -17,7 +17,7 @@ import deleteEducation from './components/controllers/cv/deleteEducation.js';
 import getSummary from './components/controllers/cv/getSummary.js';
 import postSummary from './components/controllers/cv/postSummary.js';
 
-import visitor from './components/logging/visitor.js';
+import visitor from './components/controllers/logging/visitor.js';
 import jwtMiddleware from './components/middleware/jwtMiddleware.js';
 import cookieParser from 'cookie-parser';
 import limiter from './components/middleware/rateLimitMiddleware.js';
@@ -55,7 +55,7 @@ app.delete('/deleteEducation', jwtMiddleware, deleteEducation);
 app.get('/getSummary', getSummary);
 app.post('/postSummary', jwtMiddleware, postSummary);
 
-app.post('/register', createUser);
+// app.post('/register', createUser);
 app.post('/login', login);
 app.post('/logout', logout);
 app.get('/status', jwtMiddleware, status);
